feat: adjust auth to use app router instead of pages
