--- src/pages/api/auth/[...nextauth].ts
@@ -1,29 +1,0 @@
-import NextAuth from "next-auth";
-import GoogleProvider from "next-auth/providers/google";
-
-export default NextAuth({
-    providers: [
-        GoogleProvider({
-            clientId: process.env.GOOGLE_CLIENT_ID!,
-            clientSecret: process.env.GOOGLE_CLIENT_SECRET!,
-            authorization: {
-                params: {
-                    scope: 'openid profile email https://www.googleapis.com/auth/contacts.readonly',
-                },
-            },
-        }),
-    ],
-    callbacks: {
-        async jwt({ token, account }) {
-            if (account) {
-                token.accessToken = account.access_token;
-            }
-            return token;
-        },
-        async session({ session, token }) {
-            session.accessToken = token.accessToken as string;
-            return session;
-        }
-    },
-    secret: process.env.NEXTAUTH_SECRET,
-});
